Moved ram utils into a separate file
